small change to data cleaning + fun loops
minor changes to the data cleaning code, converting the data from factor and character to numeric or date as appropriate, avoid problems with charting and incorrect data types. Converted time into rounded down integer of days passed. 

added a line to clean up unused workspace items after dataframe has been created

Also added an alternative way to create the dataframe, based on two vectors of headers and css nodes, that can be expanded or shortened as appropriate, and input parameters for how many observations are on each page, and how many pages you want. Doesn't do anything differently from what we already have. 

This is commented out at the bottom of the code, just find-replace "# " with "" and it should be good.