variables by amount/count and views
added some testing variables looking at the relationship between
different variables and the dollar amount for bribes and/or the number
of views.

Also code to read the data from github and clean it a little.

Most relevant item is probably that so many of the bribes were reported
on the same day.