ran through and commented janettes code
Ran into problems with the data formatting in the string cleaning of the population data particularly, mostly seemed to work okay, some of the hardcoded changes (deleted columns etc.) may need to be re-worked for final version if the data is a little different. 

Maybe try using full_join to combine all the scraped and produced tables together (renaming columns appropriately first), then merging and removing, mutating on that dataset.