Restructuing a bit, adding echo/eval statements
Starting to shape the main document below "blob" of code. Still doesn't
run 100%, but nearly there.

adding the "echo" statement to evaluate the segments of code but don't
print the code in the document. Moved all text down to the bottom of the
document so variables defined above can be referred to more "safely"

Code still isn't 100% stitched together